// asos_digital_wallet-f/src/app/redux/slices/settings-interface.ts
export interface ISettingsState {
  assignedView: boolean;
  logInModal: boolean;
  signUpModal: boolean;
  createTransactionModal: boolean;
  addCardModal: boolean;
  removeCardModal: boolean;
  depositBalanceModal: boolean;
  withdrawlBalanceModal: boolean;
  forgotPasswordModal: boolean;
  qrLogInModal: boolean;
  resetPasswordModal: boolean;
}